+++ example/src/main/java/com/h6ah4i/android/example/advrecyclerview/demo_d_on_longpress/DragOnLongPressExampleActivity.java
@@ -38,7 +38,7 @@ public class DragOnLongPressExampleActivity extends AppCompatActivity {
                     .add(new ExampleDataProviderFragment(), FRAGMENT_TAG_DATA_PROVIDER)
                     .commit();
             getSupportFragmentManager().beginTransaction()
-                    .add(R.id.container, new RecyclerListViewFragment(), FRAGMENT_LIST_VIEW)
+                    .add(R.id.container, new DragOnLongPressExampleFragment(), FRAGMENT_LIST_VIEW)
                     .commit();
         }
     }