+++ example/src/main/java/com/h6ah4i/android/example/advrecyclerview/demo_ds/DraggableSwipeableExampleActivity.java
@@ -44,7 +44,7 @@ public class DraggableSwipeableExampleActivity extends AppCompatActivity impleme
                     .add(new ExampleDataProviderFragment(), FRAGMENT_TAG_DATA_PROVIDER)
                     .commit();
             getSupportFragmentManager().beginTransaction()
-                    .add(R.id.container, new RecyclerListViewFragment(), FRAGMENT_LIST_VIEW)
+                    .add(R.id.container, new DraggableSwipeableExampleFragment(), FRAGMENT_LIST_VIEW)
                     .commit();
         }
     }
@@ -96,7 +96,7 @@ public class DraggableSwipeableExampleActivity extends AppCompatActivity impleme
         if (data.isPinned()) {
             // unpin if tapped the pinned item
             data.setPinned(false);
-            ((RecyclerListViewFragment) fragment).notifyItemChanged(position);
+            ((DraggableSwipeableExampleFragment) fragment).notifyItemChanged(position);
         }
     }
 
@@ -104,7 +104,7 @@ public class DraggableSwipeableExampleActivity extends AppCompatActivity impleme
         int position = getDataProvider().undoLastRemoval();
         if (position >= 0) {
             final Fragment fragment = getSupportFragmentManager().findFragmentByTag(FRAGMENT_LIST_VIEW);
-            ((RecyclerListViewFragment) fragment).notifyItemInserted(position);
+            ((DraggableSwipeableExampleFragment) fragment).notifyItemInserted(position);
         }
     }
 
@@ -114,7 +114,7 @@ public class DraggableSwipeableExampleActivity extends AppCompatActivity impleme
         final Fragment fragment = getSupportFragmentManager().findFragmentByTag(FRAGMENT_LIST_VIEW);
 
         getDataProvider().getItem(itemPosition).setPinned(ok);
-        ((RecyclerListViewFragment) fragment).notifyItemChanged(itemPosition);
+        ((DraggableSwipeableExampleFragment) fragment).notifyItemChanged(itemPosition);
     }
 
     public AbstractDataProvider getDataProvider() {