+++ example/src/main/java/com/h6ah4i/android/example/advrecyclerview/demo_e_already_expanded/AlreadyExpandedGroupsExpandableExampleActivity.java
@@ -38,7 +38,7 @@ public class AlreadyExpandedGroupsExpandableExampleActivity extends AppCompatAct
                     .add(new ExampleAddRemoveExpandableDataProviderFragment(), FRAGMENT_TAG_DATA_PROVIDER)
                     .commit();
             getSupportFragmentManager().beginTransaction()
-                    .add(R.id.container, new RecyclerListViewFragment(), FRAGMENT_LIST_VIEW)
+                    .add(R.id.container, new AlreadyExpandedGroupsExpandableExampleFragment(), FRAGMENT_LIST_VIEW)
                     .commit();
         }
     }