+++ example/src/main/java/com/h6ah4i/android/example/advrecyclerview/demo_s_viewpager/ViewPagerSwipeableExamplePagerAdapter.java
@@ -1,0 +1,52 @@
+/*
+ *    Copyright (C) 2015 Haruki Hasegawa
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package com.h6ah4i.android.example.advrecyclerview.demo_s_viewpager;
+
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentPagerAdapter;
+
+public class ViewPagerSwipeableExamplePagerAdapter extends FragmentPagerAdapter {
+
+    public ViewPagerSwipeableExamplePagerAdapter(FragmentManager fm) {
+        super(fm);
+    }
+
+    @Override
+    public Fragment getItem(int position) {
+        switch (position) {
+            case 0:
+                return ViewPagerSwipeableExampleFragment.newInstance(
+                        ViewPagerSwipeableExampleActivity.FRAGMENT_TAG_DATA_PROVIDER_1, false);
+            case 1:
+                return ViewPagerSwipeableExampleFragment.newInstance(
+                        ViewPagerSwipeableExampleActivity.FRAGMENT_TAG_DATA_PROVIDER_2, true);
+            default:
+                throw new IllegalStateException();
+        }
+    }
+
+    @Override
+    public int getCount() {
+        return 2;
+    }
+
+    @Override
+    public CharSequence getPageTitle(int position) {
+        return "Page " + (position + 1);
+    }
+}